refactoring de la page Home et PhoneBookListScreen et correction sur la page ForgotPassword
