+++ src/screen/PhoneBookDetailScreen.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { Text, View, Image, TouchableOpacity } from 'react-native';
 import SideMenu from 'react-native-side-menu';
-import Menu from './burgermenu/burgermenu';
+import Menu from './burgermenu/Menu';
 
 import { FORGOTTENPASSWORD_SCREEN_NAME } from './ForgottenPasswordScreen';
 import { HOME_SCREEN_NAME } from './HomeScreen';
